refactor: Consolidate agent routes and streamline tool handling
- Replaced agent V2 and V3 routes with a unified agent route, enhancing maintainability and reducing redundancy.
- Refactored agent tool handling to utilize a client-side console tools approach, improving responsiveness and accuracy.
- Removed deprecated files and updated imports across the codebase to reflect the new structure, ensuring consistency.
- Enhanced the Chat component to support the new agent route and tool handling, improving overall user experience.
