+++ api/src/agent-v2/index.ts
@@ -1,6 +1,6 @@
 /**
- * Agent V2 Module Exports
- * Using Vercel AI SDK for improved streaming and tool handling
+ * Agent Module Exports
+ * Using Vercel AI SDK for streaming and tool handling
  */
 
 // Types
@@ -17,14 +17,10 @@ export type {
   ReadConsoleResult,
 } from "./types";
 
-// Stream handler
-export { streamAgentResponse, processToolResult } from "./stream-handler";
-
 // Tools
-export { createUniversalToolsV2 } from "./tools/universal-tools";
+export { createUniversalTools } from "./tools/universal-tools";
 export { createSqlToolsV2 } from "./tools/sql-tools";
 export { createMongoToolsV2 } from "./tools/mongodb-tools";
-export { createConsoleToolsV2 } from "./tools/console-tools";
 
 // Prompts
 export { UNIVERSAL_PROMPT_V2 } from "./prompts/universal";
