+++ api/src/index.ts
@@ -11,8 +11,7 @@ import { databaseRoutes } from "./routes/database";
 import { dataSourceRoutes } from "./routes/sources";
 import { customPromptRoutes } from "./routes/custom-prompt";
 import { chatsRoutes } from "./routes/chats";
-import { agentV2Routes } from "./routes/agent-v2.routes";
-import { agentV3Routes } from "./routes/agent-v3.routes";
+import { agentRoutes } from "./routes/agent.routes";
 import { authRoutes } from "./auth/auth.controller";
 import { connectDatabase } from "./database/schema";
 import { workspaceRoutes } from "./routes/workspaces";
@@ -97,8 +96,7 @@ app.route("/api/workspaces/:workspaceId/flows", flowRoutes);
 app.route("/api/run", executeRoutes);
 app.route("/api/execute", executeRoutes);
 app.route("/api/database", databaseRoutes);
-app.route("/api/agent", agentV2Routes);
-app.route("/api/agent-v3", agentV3Routes);
+app.route("/api/agent", agentRoutes);
 app.route("/api/connectors", connectorRoutes);
 app.route("/api/databases", databaseSchemaRoutes);
 app.route("/api/workspaces/:workspaceId/databases", databaseTreeRoutes);
