+++ app/src/components/Chat.tsx
@@ -1,3 +1,7 @@
+/**
+ * Chat Component - Using Vercel AI SDK useChat hook
+ * Native AI SDK streaming protocol for improved compatibility
+ */
 import React, { useEffect, useMemo, useState, useRef } from "react";
 import {
   Box,
@@ -19,10 +23,9 @@ import {
   DialogTitle,
   DialogContent,
   DialogActions,
-  Switch,
-  Tooltip,
 } from "@mui/material";
 import SendIcon from "@mui/icons-material/Send";
+import StopIcon from "@mui/icons-material/Stop";
 import ReactMarkdown from "react-markdown";
 import remarkGfm from "remark-gfm";
 import { Prism as SyntaxHighlighter } from "react-syntax-highlighter";
@@ -41,27 +44,17 @@ import {
   Delete as DeleteIcon,
 } from "@mui/icons-material";
 import { useTheme as useMuiTheme } from "@mui/material/styles";
+import { useChat } from "@ai-sdk/react";
+import {
+  DefaultChatTransport,
+  lastAssistantMessageIsCompleteWithToolCalls,
+} from "ai";
 import { useWorkspace } from "../contexts/workspace-context";
 import { useConsoleStore } from "../store/consoleStore";
 import { useSettingsStore } from "../store/settingsStore";
 import { ModelSelector } from "./ModelSelector";
-
-// Note: Using simplified Message interface for this component
-interface ToolCall {
-  toolCallId?: string;
-  toolName: string;
-  timestamp?: Date | string;
-  status?: "started" | "completed";
-  // Optional fields for debugging details when available
-  input?: any;
-  result?: any;
-}
-
-interface Message {
-  role: "user" | "assistant";
-  content: string;
-  toolCalls?: Array<ToolCall>;
-}
+import { generateObjectId } from "../utils/objectId";
+import { ConsoleModification } from "../hooks/useMonacoConsole";
 
 interface ChatSessionMeta {
   _id: string;
@@ -70,6 +63,7 @@ interface ChatSessionMeta {
   updatedAt?: string;
 }
 
+// CodeBlock component for syntax highlighting
 const CodeBlock = React.memo(
   ({
     language,
@@ -88,11 +82,9 @@ const CodeBlock = React.memo(
     const [isExpanded, setIsExpanded] = React.useState(false);
     const [isCopied, setIsCopied] = React.useState(false);
 
-    // Split code into lines
     const lines = children.split("\n");
     const needsExpansion = lines.length > 12;
 
-    // Show only first 12 lines if not expanded
     const isScrollable = !!scrollable;
     const displayedCode = isScrollable
       ? children
@@ -135,7 +127,6 @@ const CodeBlock = React.memo(
             </Typography>
           </Box>
         )}
-        {/* Copy button */}
         <Box
           sx={{
             position: "absolute",
@@ -170,7 +161,7 @@ const CodeBlock = React.memo(
         </Box>
 
         <SyntaxHighlighter
-          style={syntaxTheme as any}
+          style={syntaxTheme}
           language={language}
           PreTag="div"
           customStyle={{
@@ -180,7 +171,7 @@ const CodeBlock = React.memo(
             maxWidth: "100%",
             maxHeight: isScrollable ? "50vh" : undefined,
             paddingBottom: needsExpansion && !isScrollable ? "2rem" : undefined,
-            paddingTop: "2rem", // Add padding to prevent copy button overlap
+            paddingTop: "2rem",
           }}
         >
           {displayedCode}
@@ -227,33 +218,46 @@ const CodeBlock = React.memo(
 
 CodeBlock.displayName = "CodeBlock";
 
-// ToolCallsDisplay component to show tool calls consistently
+// Tool part structure - tool type is "tool-{toolName}" with state/input/output
+interface ToolInvocationInfo {
+  toolCallId: string;
+  toolName: string;
+  state:
+    | "input-streaming"
+    | "input-available"
+    | "output-streaming"
+    | "output-available"
+    | "error";
+  input?: unknown;
+  output?: unknown;
+}
+
+// ToolCallsDisplay for showing tool invocations from AI SDK
 const ToolCallsDisplay = React.memo(
   ({
-    toolCalls,
+    toolInvocations,
     onToolClick,
   }: {
-    toolCalls?: Array<ToolCall>;
-    onToolClick?: (toolCall: ToolCall) => void;
+    toolInvocations?: ToolInvocationInfo[];
+    onToolClick?: (tool: ToolInvocationInfo) => void;
   }) => {
-    if (!toolCalls || toolCalls.length === 0) return null;
+    if (!toolInvocations || toolInvocations.length === 0) return null;
 
     return (
       <Box sx={{ mt: 1, display: "flex", flexWrap: "wrap", gap: 0.5 }}>
-        {toolCalls.map((toolCall, idx) => (
+        {toolInvocations.map(tool => (
           <Chip
-            key={
-              toolCall.toolCallId ||
-              `${toolCall.toolName}-${String(toolCall.timestamp ?? idx)}`
-            }
+            key={tool.toolCallId}
             icon={
-              toolCall.status === "completed" ? (
+              tool.state === "output-available" ? (
                 <Check sx={{ fontSize: 16 }} />
+              ) : tool.state === "error" ? (
+                <Check sx={{ fontSize: 16, color: "error.main" }} />
               ) : (
                 <CircularProgress size={14} thickness={5} />
               )
             }
-            label={toolCall.toolName}
+            label={tool.toolName}
             size="small"
             variant="outlined"
             sx={{
@@ -264,16 +268,20 @@ const ToolCallsDisplay = React.memo(
               cursor: onToolClick ? "pointer" : "default",
               "& .MuiChip-icon": {
                 color:
-                  toolCall.status === "completed"
+                  tool.state === "output-available"
                     ? "success.main"
-                    : "primary.main",
+                    : tool.state === "error"
+                      ? "error.main"
+                      : "primary.main",
               },
             }}
-            onClick={onToolClick ? () => onToolClick(toolCall) : undefined}
+            onClick={onToolClick ? () => onToolClick(tool) : undefined}
             title={
-              toolCall.status === "completed"
+              tool.state === "output-available"
                 ? "Tool executed successfully"
-                : "Tool executing..."
+                : tool.state === "error"
+                  ? "Tool execution failed"
+                  : "Tool executing..."
             }
           />
         ))}
@@ -284,323 +292,561 @@ const ToolCallsDisplay = React.memo(
 
 ToolCallsDisplay.displayName = "ToolCallsDisplay";
 
-// Rewrite MessageItem component
-const MessageItem = React.memo(
-  ({
-    message,
-    isLastUser,
-    loading,
-    onToolClick,
-  }: {
-    message: Message;
-    isLastUser: boolean;
-    loading: boolean;
-    onToolClick?: (toolCall: ToolCall) => void;
-  }) => {
-    const muiTheme = useMuiTheme();
+// ReasoningDisplay for showing reasoning/thinking parts
+const ReasoningDisplay = React.memo(
+  ({ messageParts }: { messageParts?: Array<Record<string, unknown>> }) => {
+    const [expanded, setExpanded] = React.useState(false);
 
-    // Memoize markdown rendering to keep CodeBlock stable
-    const markdownContent = React.useMemo(() => {
-      return (
-        <ReactMarkdown
-          remarkPlugins={[remarkGfm]}
-          components={{
-            code({ className, children }) {
-              const match = /language-(\w+)/.exec(className || "");
-              const isInline = !match;
-              const codeString = String(children).replace(/\n$/, "");
-              return !isInline ? (
-                <CodeBlock
-                  language={match ? match[1] : "text"}
-                  key={codeString}
-                  isGenerating={false}
-                >
-                  {codeString}
-                </CodeBlock>
-              ) : (
-                <code className={className} style={{ fontSize: "0.8rem" }}>
-                  {children}
-                </code>
-              );
-            },
-            table({ children }) {
-              return (
-                <Box sx={{ overflow: "auto", my: 1 }}>
-                  <table
-                    style={{
-                      borderCollapse: "collapse",
-                      width: "100%",
-                      fontSize: "0.875rem",
-                      border: `1px solid ${muiTheme.palette.divider}`,
-                    }}
-                  >
-                    {children}
-                  </table>
-                </Box>
-              );
-            },
-            th({ children }) {
-              return (
-                <th
-                  style={{
-                    padding: "8px 12px",
-                    textAlign: "left",
-                    backgroundColor: muiTheme.palette.background.paper,
-                    borderBottom: `2px solid ${muiTheme.palette.divider}`,
-                    borderRight: `1px solid ${muiTheme.palette.divider}`,
-                    fontWeight: 600,
-                  }}
-                >
-                  {children}
-                </th>
-              );
-            },
-            td({ children }) {
-              return (
-                <td
-                  style={{
-                    padding: "8px 12px",
-                    borderBottom: `1px solid ${muiTheme.palette.divider}`,
-                    borderRight: `1px solid ${muiTheme.palette.divider}`,
-                    backgroundColor: muiTheme.palette.background.paper,
-                  }}
-                >
-                  {children}
-                </td>
-              );
-            },
-          }}
-        >
-          {message.content}
-        </ReactMarkdown>
-      );
-    }, [
-      message.content,
-      muiTheme.palette.divider,
-      muiTheme.palette.background.paper,
-    ]);
+    if (!messageParts) return null;
+
+    const reasoningParts = messageParts.filter(
+      (p): p is { type: "reasoning"; text: string } =>
+        p.type === "reasoning" && typeof p.text === "string",
+    );
+
+    if (reasoningParts.length === 0) return null;
 
     return (
-      <ListItem alignItems="flex-start" sx={{ p: 0 }}>
-        {message.role === "user" ? (
-          <Box sx={{ flex: 1 }}>
-            <Paper
-              variant="outlined"
-              sx={{
-                p: 1,
-                borderRadius: 1,
-                backgroundColor: "background.paper",
-                overflow: "hidden",
-              }}
-            >
-              <Box
-                sx={{
-                  overflow: "auto",
-                  maxWidth: "100%",
-                }}
-              >
-                <ListItemText
-                  primary={message.content}
-                  primaryTypographyProps={{
-                    variant: "body2",
-                    color: "text.primary",
-                    sx: {
-                      whiteSpace: "pre-wrap",
-                      wordBreak: "break-word",
-                      overflowWrap: "break-word",
-                    },
-                  }}
-                />
-              </Box>
-              {/* Generating notice inside last user message */}
-              {isLastUser && loading && (
-                <Typography
-                  variant="body2"
-                  sx={{ color: "text.secondary", mt: 0.5 }}
-                >
-                  Generating...
-                </Typography>
-              )}
-            </Paper>
-          </Box>
-        ) : (
+      <Box sx={{ my: 1 }}>
+        <Button
+          size="small"
+          onClick={() => setExpanded(!expanded)}
+          startIcon={expanded ? <ExpandLess /> : <ExpandMore />}
+          sx={{
+            color: "text.secondary",
+            textTransform: "none",
+            fontSize: "0.8rem",
+            p: 0,
+            minWidth: "auto",
+            "&:hover": {
+              backgroundColor: "transparent",
+              textDecoration: "underline",
+            },
+          }}
+          disableRipple
+        >
+          Thinking Process
+        </Button>
+        {expanded && (
           <Box
             sx={{
-              flex: 1,
-              overflow: "hidden",
+              mt: 1,
+              pl: 2,
+              borderLeft: 2,
+              borderColor: "divider",
+              color: "text.secondary",
               fontSize: "0.875rem",
-
-              "& pre": { margin: 0, overflow: "hidden" },
             }}
           >
-            {/* Display tool calls first (chronologically before response) */}
-            {message.role === "assistant" && (
-              <ToolCallsDisplay
-                toolCalls={message.toolCalls}
-                onToolClick={onToolClick}
-              />
-            )}
-            {/* Then display the message content */}
-            {markdownContent}
+            {reasoningParts.map((part, i) => (
+              <Box key={i} sx={{ mb: 1, whiteSpace: "pre-wrap" }}>
+                {part.text}
+              </Box>
+            ))}
           </Box>
         )}
-      </ListItem>
+      </Box>
     );
   },
 );
 
-MessageItem.displayName = "MessageItem";
+ReasoningDisplay.displayName = "ReasoningDisplay";
+
+// Extended ConsoleModification with fields for console creation
+type ConsoleModificationPayload = ConsoleModification & {
+  consoleId?: string;
+  title?: string;
+  connectionId?: string;
+  databaseId?: string;
+  databaseName?: string;
+};
 
 interface ChatProps {
-  onConsoleModification?: (modification: any) => void;
+  onConsoleModification?: (modification: ConsoleModificationPayload) => void;
 }
 
 const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
+  const muiTheme = useMuiTheme();
   const { currentWorkspace } = useWorkspace();
   const selectedModelId = useSettingsStore(s => s.selectedModelId);
-  const useChatV3 = useSettingsStore(s => s.useChatV3);
-  const setUseChatV3 = useSettingsStore(s => s.setUseChatV3);
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [input, setInput] = useState("");
-  const [loading, setLoading] = useState(false);
-  const [sessions, setSessions] = useState<ChatSessionMeta[]>([]);
-  const [sessionId, setSessionId] = useState<string | "">("");
-  const [steps, setSteps] = useState<string[]>([]);
-  const [streamingContent, setStreamingContent] = useState<string>("");
-  const [error, setError] = useState<string | null>(null);
-  const [streamingToolCalls, setStreamingToolCalls] = useState<
-    Array<{
-      toolCallId: string;
-      toolName: string;
-      timestamp: string;
-      status: "started" | "completed";
-      input?: any;
-      result?: any;
-    }>
-  >([]);
-  const streamingToolCallsRef = useRef<
-    Array<{
-      toolCallId: string;
-      toolName: string;
-      timestamp: string;
-      status: "started" | "completed";
-      input?: any;
-      result?: any;
-    }>
-  >([]);
-  const [toolDialogOpen, setToolDialogOpen] = useState(false);
-  const [selectedToolCall, setSelectedToolCall] = useState<ToolCall | null>(
-    null,
-  );
-
-  // Get console store
   const { consoleTabs, activeConsoleId } = useConsoleStore();
 
-  // Only real console tabs count as "consoles" for the agent.
-  // Tabs like Settings/Connectors/etc must never be sent as console context.
-  const realConsoleTabs = useMemo(
-    () =>
-      (consoleTabs || []).filter(
-        (t: any) => t?.kind === undefined || t?.kind === "console",
-      ),
-    [consoleTabs],
-  );
-
-  // History menu state
+  const [sessions, setSessions] = useState<ChatSessionMeta[]>([]);
+  // chatId is a MongoDB ObjectId generated locally - frontend owns the ID (AI SDK best practice)
+  const [chatId, setChatId] = useState<string>(() => generateObjectId());
   const [historyMenuAnchor, setHistoryMenuAnchor] =
     useState<null | HTMLElement>(null);
   const historyMenuOpen = Boolean(historyMenuAnchor);
-
-  // Ref for auto-focusing the input
   const inputRef = useRef<HTMLInputElement>(null);
+  const messagesEndRef = useRef<HTMLDivElement>(null);
 
-  // ---------------------------------------------------------------------------
-  // Session management (identical to Chat2)
-  // ---------------------------------------------------------------------------
+  // Track if we're viewing an existing chat from history (vs a new chat)
+  // Moved before useChat so onFinish callback can access it
+  const [isExistingChat, setIsExistingChat] = useState(false);
 
-  useEffect(() => {
-    const fetchSessions = async () => {
-      if (!currentWorkspace) return;
+  // Refs for accessing current values in callbacks (avoids stale closures)
+  const isExistingChatRef = useRef(isExistingChat);
+  isExistingChatRef.current = isExistingChat;
 
-      try {
-        const res = await fetch(`/api/workspaces/${currentWorkspace.id}/chats`);
-        if (res.ok) {
-          const data = await res.json();
-          setSessions(data);
-          // Use functional update to avoid sessionId dependency
-          if (data.length > 0) {
-            setSessionId(current => (current ? current : data[0]._id));
-          }
-        }
-      } catch (_) {
-        /* ignore */
-      }
-    };
-    fetchSessions();
-  }, [currentWorkspace]);
+  // Ref for onConsoleModification to avoid stale closure in onToolCall
+  const onConsoleModificationRef = useRef(onConsoleModification);
+  onConsoleModificationRef.current = onConsoleModification;
 
-  useEffect(() => {
-    const loadSession = async () => {
-      if (!sessionId || !currentWorkspace) {
-        setMessages([]);
-        return;
-      }
-      try {
-        const res = await fetch(
-          `/api/workspaces/${currentWorkspace.id}/chats/${sessionId}`,
-        );
-        if (res.ok) {
-          const data = await res.json();
-          setMessages(data.messages || []);
-        }
-      } catch (_) {
-        /* ignore */
-      }
-    };
-    loadSession();
-  }, [sessionId, currentWorkspace]);
-
-  // Auto-focus input when session changes or when creating new chat
-  useEffect(() => {
-    // Focus after a short delay to ensure the component is rendered
-    const timer = setTimeout(() => {
-      if (inputRef.current) {
-        inputRef.current.focus();
-      }
-    }, 100);
-
-    return () => clearTimeout(timer);
-  }, [sessionId, messages.length]);
-
-  const createNewSession = async () => {
+  // Function to fetch sessions - defined before useChat so it can be used in onFinish
+  // Using a ref-based pattern to always access the current workspace
+  const fetchSessionsRef = useRef<() => Promise<void>>();
+  fetchSessionsRef.current = async () => {
     if (!currentWorkspace) return;
-
     try {
-      const res = await fetch(`/api/workspaces/${currentWorkspace.id}/chats`, {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ title: "New Chat" }),
-      });
+      const res = await fetch(`/api/workspaces/${currentWorkspace.id}/chats`);
       if (res.ok) {
         const data = await res.json();
-        const newId = data.chatId as string;
-        // Refresh sessions list
-        const sessionsRes = await fetch(
-          `/api/workspaces/${currentWorkspace.id}/chats`,
-        );
-        if (sessionsRes.ok) {
-          const sessionsData = await sessionsRes.json();
-          setSessions(sessionsData);
-        }
-        setSessionId(newId);
-        setMessages([]);
+        setSessions(data);
       }
-    } catch (_) {
+    } catch {
       /* ignore */
     }
   };
 
-  // History menu handlers
+  // Tool debug dialog
+  const [toolDialogOpen, setToolDialogOpen] = useState(false);
+  const [selectedTool, setSelectedTool] = useState<ToolInvocationInfo | null>(
+    null,
+  );
+
+  // Filter to only real console tabs
+  const realConsoleTabs = useMemo(
+    () =>
+      (consoleTabs || []).filter(
+        t => t?.kind === undefined || t?.kind === "console",
+      ),
+    [consoleTabs],
+  );
+
+  // Build consoles data for the backend
+  const consolesData = useMemo(() => {
+    const data: Array<{
+      id: string;
+      title: string;
+      content: string;
+      connectionId?: string;
+      databaseId?: string;
+      databaseName?: string;
+    }> = [];
+
+    for (const tab of realConsoleTabs) {
+      if (!tab?.id) continue;
+      const connectionId = tab?.connectionId;
+      const databaseId =
+        tab?.databaseId || tab?.metadata?.queryOptions?.databaseId;
+      const databaseName =
+        tab?.databaseName ||
+        tab?.metadata?.queryOptions?.databaseName ||
+        tab?.metadata?.queryOptions?.dbName;
+
+      data.push({
+        id: tab.id,
+        title: tab.title,
+        content: tab.content || "",
+        connectionId,
+        databaseId,
+        databaseName,
+      });
+    }
+    return data;
+  }, [realConsoleTabs]);
+
+  // Local input state
+  const [input, setInput] = useState("");
+
+  // Create transport with dynamic body based on current state
+  const transport = useMemo(
+    () =>
+      new DefaultChatTransport({
+        api: "/api/agent/chat",
+        body: {
+          workspaceId: currentWorkspace?.id,
+          consoles: consolesData,
+          consoleId: activeConsoleId,
+          modelId: selectedModelId,
+          chatId, // Frontend-owned ID (AI SDK best practice)
+        },
+      }),
+    [
+      currentWorkspace?.id,
+      consolesData,
+      activeConsoleId,
+      selectedModelId,
+      chatId,
+    ],
+  );
+
+  // Note: We use (useConsoleStore as any).getState() inside callbacks to avoid stale closure issues
+
+  // useChat hook from Vercel AI SDK
+  // IMPORTANT: The 'id' prop is critical - it resets the hook's internal message state
+  // when chatId changes. Without it, switching chats causes stale messages to persist.
+  // @typescript-eslint/no-explicit-any
+  const {
+    messages,
+    sendMessage,
+    status,
+    error,
+    stop,
+    setMessages,
+    addToolOutput,
+  } = useChat({
+    id: chatId, // Reset hook state when chatId changes (fixes stale messages bug)
+    transport: transport as any, // Type assertion to handle pnpm version resolution
+
+    // Automatically submit when all tool results are available
+    sendAutomaticallyWhen: lastAssistantMessageIsCompleteWithToolCalls,
+
+    // Handle client-side tools (console operations)
+    async onToolCall({ toolCall }) {
+      // Skip dynamic tools (not our console tools)
+      if ((toolCall as { dynamic?: boolean }).dynamic) {
+        return;
+      }
+
+      const toolName = toolCall.toolName;
+      const input = toolCall.input as Record<string, unknown>;
+
+      // Handle read_console
+      if (toolName === "read_console") {
+        // Get fresh state to avoid stale closure issues
+        const currentStore = (useConsoleStore as any).getState();
+        const currentTabs = currentStore.consoleTabs;
+        const currentActiveId = currentStore.activeConsoleId;
+
+        const consoleId = (input.consoleId as string | null) ?? currentActiveId;
+        const targetConsole = consoleId
+          ? currentTabs.find((c: any) => c.id === consoleId)
+          : currentTabs.find((c: any) => c.id === currentActiveId) ||
+            currentTabs[0];
+
+        if (!targetConsole) {
+          addToolOutput({
+            tool: "read_console",
+            toolCallId: toolCall.toolCallId,
+            output: {
+              success: false,
+              error: consoleId
+                ? `Console with ID ${consoleId} not found`
+                : "No console is currently active",
+            },
+          });
+          return;
+        }
+
+        addToolOutput({
+          tool: "read_console",
+          toolCallId: toolCall.toolCallId,
+          output: {
+            success: true,
+            consoleId: targetConsole.id,
+            title: targetConsole.title,
+            content: targetConsole.content || "",
+            connectionId: targetConsole.connectionId,
+            connectionType: (
+              targetConsole.metadata as { connectionType?: string }
+            )?.connectionType,
+            databaseId: targetConsole.databaseId,
+            databaseName: targetConsole.databaseName,
+          },
+        });
+        return;
+      }
+
+      // Handle modify_console - dispatch through event system for proper Monaco update
+      if (toolName === "modify_console") {
+        // Get fresh state to avoid stale closure issues
+        const currentStore = (useConsoleStore as any).getState();
+        const currentTabs = currentStore.consoleTabs;
+        const currentActiveId = currentStore.activeConsoleId;
+
+        const action = input.action as "replace" | "insert" | "append";
+        const content = input.content as string;
+        const position = input.position as number | null;
+        const inputConsoleId = input.consoleId as string | null | undefined;
+
+        // Determine target console
+        const resolvedConsoleId =
+          inputConsoleId ?? currentActiveId ?? currentTabs[0]?.id;
+
+        if (!resolvedConsoleId) {
+          addToolOutput({
+            tool: "modify_console",
+            toolCallId: toolCall.toolCallId,
+            output: {
+              success: false,
+              error: "No console is currently open. Call create_console first.",
+            },
+          });
+          return;
+        }
+
+        const targetConsole = currentTabs.find(
+          (c: any) => c.id === resolvedConsoleId,
+        );
+        if (!targetConsole) {
+          addToolOutput({
+            tool: "modify_console",
+            toolCallId: toolCall.toolCallId,
+            output: {
+              success: false,
+              error: `Console with ID ${resolvedConsoleId} not found`,
+            },
+          });
+          return;
+        }
+
+        // Validate insert action has position
+        if (
+          action === "insert" &&
+          (position === null || position === undefined)
+        ) {
+          addToolOutput({
+            tool: "modify_console",
+            toolCallId: toolCall.toolCallId,
+            output: {
+              success: false,
+              error: "Position is required for insert action",
+            },
+          });
+          return;
+        }
+
+        // Dispatch through the event system - this ensures Monaco editor gets updated
+        // The App.tsx handleConsoleModification callback will:
+        // 1. Dispatch a CustomEvent that Editor.tsx listens to
+        // 2. Editor.tsx calls showDiff() on the Console ref
+        // 3. Console.tsx updates Monaco editor via the diff mode
+        if (onConsoleModificationRef.current) {
+          onConsoleModificationRef.current({
+            action,
+            content,
+            // Convert line number to position format expected by ConsoleModification
+            position:
+              position !== null && position !== undefined
+                ? { line: position, column: 1 }
+                : undefined,
+            consoleId: resolvedConsoleId,
+          });
+        }
+
+        // Also update store for consistency
+        const currentContent = targetConsole.content || "";
+        let newContent: string;
+
+        switch (action) {
+          case "replace":
+            newContent = content;
+            break;
+          case "append":
+            newContent =
+              currentContent +
+              (currentContent.endsWith("\n") ? "" : "\n") +
+              content;
+            break;
+          case "insert":
+            if (position !== null && position !== undefined) {
+              const lines = currentContent.split("\n");
+              const insertIndex = Math.max(0, position - 1);
+              lines.splice(insertIndex, 0, content);
+              newContent = lines.join("\n");
+            } else {
+              newContent = content + currentContent;
+            }
+            break;
+          default:
+            newContent = content;
+        }
+        currentStore.updateConsoleContent(resolvedConsoleId, newContent);
+
+        addToolOutput({
+          tool: "modify_console",
+          toolCallId: toolCall.toolCallId,
+          output: {
+            success: true,
+            consoleId: resolvedConsoleId,
+            message: `Console ${action}d successfully`,
+          },
+        });
+        return;
+      }
+
+      // Handle create_console - dispatch through event system
+      if (toolName === "create_console") {
+        // Get fresh state to avoid stale closure issues
+        const currentStore = (useConsoleStore as any).getState();
+        const currentTabs = currentStore.consoleTabs;
+        const currentActiveId = currentStore.activeConsoleId;
+
+        const title = input.title as string;
+        const content = input.content as string;
+        const connectionId = (input.connectionId as string | null) ?? undefined;
+        const databaseId = (input.databaseId as string | null) ?? undefined;
+        const databaseName = (input.databaseName as string | null) ?? undefined;
+
+        // If connection info not provided, inherit from active console
+        const baseConsole =
+          currentTabs.find((c: any) => c.id === currentActiveId) ||
+          currentTabs[0];
+
+        const effectiveConnectionId = connectionId ?? baseConsole?.connectionId;
+        const effectiveDatabaseId = databaseId ?? baseConsole?.databaseId;
+        const effectiveDatabaseName = databaseName ?? baseConsole?.databaseName;
+
+        // Generate a new ID for the console
+        const newConsoleId = generateObjectId();
+
+        // Dispatch through the event system - App.tsx handleConsoleModification will:
+        // 1. Call addConsoleTab with the provided consoleId
+        // 2. Call setActiveConsole
+        if (onConsoleModificationRef.current) {
+          onConsoleModificationRef.current({
+            action: "create",
+            content,
+            consoleId: newConsoleId,
+            title,
+            connectionId: effectiveConnectionId,
+            databaseId: effectiveDatabaseId,
+            databaseName: effectiveDatabaseName,
+          });
+        }
+
+        addToolOutput({
+          tool: "create_console",
+          toolCallId: toolCall.toolCallId,
+          output: {
+            success: true,
+            _eventType: "console_creation",
+            consoleId: newConsoleId,
+            title,
+            content,
+            connectionId: effectiveConnectionId,
+            databaseId: effectiveDatabaseId,
+            databaseName: effectiveDatabaseName,
+            message: `✓ New console "${title}" created successfully`,
+          },
+        });
+        return;
+      }
+
+      // Unknown tool - not a client-side console tool, let it be handled server-side
+    },
+
+    onError: err => {
+      console.error("[Chat] Error:", err);
+    },
+    onFinish: () => {
+      // When a new chat's first message exchange completes, refresh the sessions list
+      // so the newly saved chat appears in the history menu
+      if (!isExistingChatRef.current) {
+        fetchSessionsRef.current?.();
+      }
+    },
+  });
+
+  const isLoading = status === "streaming" || status === "submitted";
+
+  // Auto-scroll to bottom when messages change
+  useEffect(() => {
+    messagesEndRef.current?.scrollIntoView({ behavior: "smooth" });
+  }, [messages]);
+
+  // Session management - fetch available chat sessions for history menu
+  useEffect(() => {
+    fetchSessionsRef.current?.();
+  }, [currentWorkspace]);
+
+  // Load messages when selecting an existing chat from history
+  useEffect(() => {
+    const loadSession = async () => {
+      if (!isExistingChat || !currentWorkspace) {
+        return;
+      }
+      try {
+        const res = await fetch(
+          `/api/workspaces/${currentWorkspace.id}/chats/${chatId}`,
+        );
+        if (res.ok) {
+          const data = await res.json();
+          // Convert stored messages to AI SDK format with parts including tool calls
+          const convertedMessages =
+            data.messages?.map((msg: any) => {
+              const parts: Array<Record<string, unknown>> = [];
+
+              // Add tool call parts first (they execute before text response)
+              // IMPORTANT: input must always be defined (at least {}) for OpenAI API compatibility
+              // The API requires 'arguments' field which comes from 'input'
+              if (msg.toolCalls && msg.toolCalls.length > 0) {
+                for (const tc of msg.toolCalls) {
+                  // Skip tool calls without a valid toolName
+                  if (!tc.toolName) continue;
+
+                  parts.push({
+                    type: `tool-${tc.toolName}`,
+                    toolCallId:
+                      tc.toolCallId ||
+                      tc._id?.toString() ||
+                      `saved-${tc.toolName}-${Date.now()}-${Math.random()}`,
+                    toolName: tc.toolName,
+                    state: "output-available",
+                    // CRITICAL: input must never be undefined - OpenAI API requires 'arguments'
+                    input: tc.input ?? {},
+                    output: tc.result ?? null,
+                  });
+                }
+              }
+
+              // Add reasoning parts (if any)
+              if (msg.reasoning && Array.isArray(msg.reasoning)) {
+                for (const reasoningText of msg.reasoning) {
+                  parts.push({
+                    type: "reasoning",
+                    text: reasoningText,
+                  });
+                }
+              }
+
+              // Add text content part
+              if (msg.content) {
+                parts.push({ type: "text", text: msg.content });
+              }
+
+              return {
+                id:
+                  msg._id?.toString() ||
+                  msg.id ||
+                  `${Date.now()}-${Math.random()}`,
+                role: msg.role,
+                parts,
+              };
+            }) || [];
+          setMessages(convertedMessages);
+        }
+      } catch {
+        /* ignore */
+      }
+    };
+    loadSession();
+  }, [chatId, isExistingChat, currentWorkspace, setMessages]);
+
+  // Focus input
+  useEffect(() => {
+    const timer = setTimeout(() => inputRef.current?.focus(), 100);
+    return () => clearTimeout(timer);
+  }, [chatId, messages.length]);
+
+  // Create new chat session - just generate a new ID locally (no API call needed)
+  const createNewSession = () => {
+    setChatId(generateObjectId());
+    setMessages([]);
+    setIsExistingChat(false);
+  };
+
   const handleHistoryMenuOpen = (event: React.MouseEvent<HTMLElement>) => {
     setHistoryMenuAnchor(event.currentTarget);
   };
@@ -609,422 +855,165 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
     setHistoryMenuAnchor(null);
   };
 
-  const handleSelectSession = (sessionId: string) => {
-    setSessionId(sessionId);
+  const handleSelectSession = (id: string) => {
+    setChatId(id);
+    setMessages([]);
+    setIsExistingChat(true);
     handleHistoryMenuClose();
   };
 
-  const handleDeleteSession = async (
-    sessionIdToDelete: string,
-    event: React.MouseEvent,
-  ) => {
-    event.stopPropagation();
+  const handleDeleteSession = async (id: string, e: React.MouseEvent) => {
+    e.stopPropagation();
     if (!currentWorkspace) return;
-
     try {
       const res = await fetch(
-        `/api/workspaces/${currentWorkspace.id}/chats/${sessionIdToDelete}`,
-        {
-          method: "DELETE",
-        },
+        `/api/workspaces/${currentWorkspace.id}/chats/${id}`,
+        { method: "DELETE" },
       );
       if (res.ok) {
-        // Refresh sessions list
-        const sessionsRes = await fetch(
-          `/api/workspaces/${currentWorkspace.id}/chats`,
-        );
-        if (sessionsRes.ok) {
-          const sessionsData = await sessionsRes.json();
-          setSessions(sessionsData);
-          // If we deleted the current session, switch to another one or create new
-          if (sessionIdToDelete === sessionId) {
-            if (sessionsData.length > 0) {
-              setSessionId(sessionsData[0]._id);
-            } else {
-              createNewSession();
-            }
-          }
+        const newSessions = sessions.filter(s => s._id !== id);
+        setSessions(newSessions);
+        if (chatId === id) {
+          // If we deleted the current chat, start a new one
+          createNewSession();
         }
       }
-    } catch (_) {
+    } catch {
       /* ignore */
     }
   };
 
-  // ---------------------------------------------------------------------------
-  // Messaging helpers
-  // ---------------------------------------------------------------------------
-
-  const streamResponse = async (latestMessage: string) => {
-    if (!currentWorkspace) {
-      throw new Error("No workspace selected");
-    }
-
-    // Helper to extract database context from console tab
-    const extractDatabaseContext = (tab: any) => {
-      // connectionId = DatabaseConnection MongoDB ObjectId
-      const connectionId = tab?.connectionId;
-      // databaseId = specific database ID (e.g., D1 UUID for cluster mode)
-      const databaseId =
-        tab?.databaseId || tab?.metadata?.queryOptions?.databaseId;
-      // databaseName = human-readable database name
-      const databaseName =
-        tab?.databaseName ||
-        tab?.metadata?.queryOptions?.databaseName ||
-        tab?.metadata?.queryOptions?.dbName;
-      return { connectionId, databaseId, databaseName };
-    };
-
-    // Prepare consoles data for the backend
-    const consolesData: Array<{
-      id: string;
-      title: string;
-      content: string;
-      connectionId?: string;
-      databaseId?: string;
-      databaseName?: string;
-      metadata?: Record<string, any>;
-    }> = [];
-
-    // Include the active console (only if it's a real console tab)
-    if (activeConsoleId && realConsoleTabs.length > 0) {
-      const activeConsole = realConsoleTabs.find(
-        tab => tab.id === activeConsoleId,
-      );
-
-      if (activeConsole) {
-        const { connectionId, databaseId, databaseName } =
-          extractDatabaseContext(activeConsole);
-
-        consolesData.push({
-          id: activeConsoleId,
-          title: activeConsole.title,
-          content: activeConsole.content || "",
-          connectionId,
-          databaseId,
-          databaseName,
-          metadata: {
-            connectionId,
-            databaseId,
-            databaseName,
-            queryOptions: activeConsole.metadata?.queryOptions,
-          },
-        });
-      }
-    }
-
-    // Include other open console tabs (metadata + content) so the backend can pick a valid
-    // console even when the user is focused on a non-console tab (e.g., Settings).
-    for (const tab of realConsoleTabs) {
-      if (!tab?.id) continue;
-      if (consolesData.some(c => c.id === tab.id)) continue;
-
-      const { connectionId, databaseId, databaseName } =
-        extractDatabaseContext(tab);
-
-      consolesData.push({
-        id: tab.id,
-        title: tab.title,
-        content: tab.content || "",
-        connectionId,
-        databaseId,
-        databaseName,
-        metadata: {
-          connectionId,
-          databaseId,
-          databaseName,
-          queryOptions: tab.metadata?.queryOptions,
-        },
-      });
-    }
-
-    const response = await fetch("/api/agent/stream", {
-      method: "POST",
-      headers: { "Content-Type": "application/json" },
-      body: JSON.stringify({
-        sessionId,
-        message: latestMessage,
-        workspaceId: currentWorkspace.id,
-        consoles: consolesData,
-        modelId: selectedModelId,
-      }),
-    });
-
-    if (!response.ok) {
-      const errorText = await response.text();
-      let errorMessage = `HTTP ${response.status}`;
-      try {
-        const errorJson = JSON.parse(errorText);
-        errorMessage = errorJson.error || errorJson.message || errorMessage;
-      } catch {
-        // If not JSON, use the text directly
-        if (errorText) {
-          errorMessage = errorText;
-        }
-      }
-      throw new Error(errorMessage);
-    }
-
-    if (!response.body) {
-      throw new Error("No response body");
-    }
-
-    const reader = response.body.getReader();
-    const decoder = new TextDecoder();
-    let assistantContent = "";
-    let done = false;
-
-    // Don't add optimistic message while loading, handle streaming separately
-    setStreamingContent("");
-    setError(null); // Clear any previous errors
-    setStreamingToolCalls([]); // Clear tool calls from previous message
-    streamingToolCallsRef.current = [];
-
-    while (!done) {
-      const { value, done: doneReading } = await reader.read();
-      done = doneReading;
-      const chunkValue = decoder.decode(value || new Uint8Array());
-
-      const lines = chunkValue.split("\n\n").filter(Boolean);
-      for (const line of lines) {
-        if (line.startsWith("data: ")) {
-          const data = line.replace("data: ", "");
-          if (data === "[DONE]") {
-            done = true;
-            break;
-          }
-
-          try {
-            const parsed = JSON.parse(data);
-
-            // Handle different event types
-            if (parsed.type === "text") {
-              assistantContent += parsed.content;
-              setStreamingContent(assistantContent);
-            } else if (parsed.type === "step" && parsed.name) {
-              setSteps(prev => [...prev, parsed.name]);
-
-              // Track tool calls from step events
-              if (parsed.name.startsWith("tool_called:")) {
-                const toolName = parsed.name.replace("tool_called:", "");
-                const toolCallId =
-                  parsed.toolCallId ||
-                  `${toolName}-${Date.now()}-${Math.random().toString(16).slice(2)}`;
-                setStreamingToolCalls(prev => {
-                  const existing = prev.find(
-                    tc => tc.toolCallId === toolCallId,
-                  );
-                  const updated = existing
-                    ? prev.map(tc =>
-                        tc.toolCallId === toolCallId
-                          ? {
-                              ...tc,
-                              status: "started" as const,
-                              timestamp: new Date().toISOString(),
-                              input:
-                                parsed.input ??
-                                parsed.args ??
-                                parsed.parameters ??
-                                parsed.payload ??
-                                tc.input,
-                            }
-                          : tc,
-                      )
-                    : [
-                        ...prev,
-                        {
-                          toolCallId,
-                          toolName,
-                          timestamp: new Date().toISOString(),
-                          status: "started" as const,
-                          input:
-                            parsed.input ??
-                            parsed.args ??
-                            parsed.parameters ??
-                            parsed.payload,
-                        },
-                      ];
-                  streamingToolCallsRef.current = updated;
-                  return updated;
-                });
-              } else if (parsed.name.startsWith("tool_output:")) {
-                const toolName = parsed.name.replace("tool_output:", "");
-                const toolCallId = parsed.toolCallId as string | undefined;
-                setStreamingToolCalls(prev => {
-                  const maybeResult =
-                    parsed.output ?? parsed.result ?? parsed.data;
-                  let updated: typeof prev;
-                  if (toolCallId) {
-                    updated = prev.map(tc =>
-                      tc.toolCallId === toolCallId
-                        ? {
-                            ...tc,
-                            status: "completed" as const,
-                            result: maybeResult ?? tc.result,
-                          }
-                        : tc,
-                    );
-                  } else {
-                    const reverseIndex = [...prev]
-                      .reverse()
-                      .findIndex(
-                        tc =>
-                          tc.toolName === toolName && tc.status === "started",
-                      );
-                    if (reverseIndex === -1) {
-                      updated = prev;
-                    } else {
-                      const targetIndex = prev.length - 1 - reverseIndex;
-                      updated = prev.map((tc, idx) =>
-                        idx === targetIndex
-                          ? {
-                              ...tc,
-                              status: "completed" as const,
-                              result: maybeResult ?? tc.result,
-                            }
-                          : tc,
-                      );
-                    }
-                  }
-                  streamingToolCallsRef.current = updated;
-                  return updated;
-                });
-              }
-            } else if (
-              parsed.type === "session" &&
-              parsed.sessionId &&
-              !sessionId
-            ) {
-              setSessionId(parsed.sessionId);
-            } else if (
-              parsed.type === "console_modification" &&
-              parsed.modification
-            ) {
-              if (onConsoleModification) {
-                // Pass the modification with the consoleId if available
-                onConsoleModification({
-                  ...parsed.modification,
-                  consoleId: parsed.consoleId,
-                });
-              }
-            } else if (parsed.type === "console_creation") {
-              // Handle console creation event
-              if (onConsoleModification) {
-                // Use the modification handler to create a new console
-                onConsoleModification({
-                  action: "create",
-                  content: parsed.content || "",
-                  consoleId: parsed.consoleId,
-                  title: parsed.title,
-                  connectionId: parsed.connectionId,
-                  databaseId: parsed.databaseId,
-                  databaseName: parsed.databaseName,
-                });
-              }
-            } else if (parsed.type === "handoff") {
-              // Handle handoff events - show a status message but don't save it
-              setSteps(prev => [
-                ...prev,
-                `Switching to ${parsed.agent} assistant`,
-              ]);
-            } else if (parsed.type === "error") {
-              // Handle error events
-              console.error("Error from agent:", parsed.message);
-              setError(parsed.message || "An error occurred");
-              // Still add the error to messages so it's visible in chat history
-              assistantContent = `Error: ${parsed.message || "An unknown error occurred"}`;
-              setStreamingContent(assistantContent);
-            }
-          } catch (parseError) {
-            // Track 1.2: Surface parse errors instead of silent ignore
-            console.error(
-              "[Chat] SSE parse error:",
-              parseError,
-              "Raw data:",
-              data,
-            );
-            // Surface critical failures for console modifications
-            if (
-              data.includes("console_modification") ||
-              data.includes("console_creation")
-            ) {
-              setError(
-                "Failed to apply console changes. The AI response may be incomplete.",
-              );
-            }
-          }
-        }
-      }
-    }
-
-    // After streaming is complete, add the final message to the messages array
-    const latestToolCalls = streamingToolCallsRef.current;
-    if (assistantContent || latestToolCalls.length > 0) {
-      // Capture tool calls before clearing using ref to avoid stale closure
-      const finalToolCalls =
-        latestToolCalls.length > 0 ? [...latestToolCalls] : undefined;
-
-      setMessages(prev => [
-        ...prev,
-        {
-          role: "assistant",
-          content: assistantContent || "",
-          toolCalls: finalToolCalls,
-        },
-      ]);
-
-      // Clear states after adding message
-      setSteps([]);
-      setStreamingContent("");
-      setStreamingToolCalls([]);
-      streamingToolCallsRef.current = [];
-    }
-  };
-
-  const sendMessage = async () => {
-    if (!input.trim() || !currentWorkspace) return;
-
-    const userMessage = input.trim();
-
-    // Clear any previous errors
-    setError(null);
-
-    // Optimistically add user message
-    setMessages(prev => [...prev, { role: "user", content: userMessage }]);
-    setSteps([]);
-    setInput("");
-    setLoading(true);
-
-    try {
-      await streamResponse(userMessage);
-    } catch (err: any) {
-      setError(err.message || "Failed to send message");
-      setMessages(prev => [
-        ...prev,
-        {
-          role: "assistant",
-          content: `Error: ${err.message}`,
-        },
-      ]);
-    } finally {
-      setLoading(false);
-    }
-  };
-
   // Tool debug dialog handlers
-  const handleToolClick = (toolCall: ToolCall) => {
-    setSelectedToolCall(toolCall);
+  const handleToolClick = (tool: ToolInvocationInfo) => {
+    setSelectedTool(tool);
     setToolDialogOpen(true);
   };
 
   const handleCloseToolDialog = () => {
     setToolDialogOpen(false);
-    setSelectedToolCall(null);
+    setSelectedTool(null);
   };
 
-  // ---------------------------------------------------------------------------
-  // Render
-  // ---------------------------------------------------------------------------
+  // Extract tool invocations from message parts
+  // AI SDK has two tool part types:
+  // - Static tools: type is "tool-{toolName}" (e.g., "tool-list_connections")
+  // - Dynamic tools: type is "dynamic-tool" with toolName as separate property
+  const getToolInvocations = (
+    messageParts: Array<Record<string, unknown>> | undefined,
+  ): ToolInvocationInfo[] => {
+    if (!messageParts) return [];
+    return messageParts
+      .filter(part => {
+        const type = part.type;
+        if (typeof type !== "string") return false;
+        // Match static tools (type starts with "tool-") or dynamic tools
+        return type.startsWith("tool-") || type === "dynamic-tool";
+      })
+      .map(part => {
+        const partType = part.type as string;
+        // For dynamic tools, use the toolName property; for static tools, extract from type
+        // Static tool names: "tool-{name}" -> split on "-" and rejoin (handles names with hyphens)
+        const toolName =
+          partType === "dynamic-tool"
+            ? (part.toolName as string)
+            : partType.split("-").slice(1).join("-");
+        return {
+          toolCallId: (part.toolCallId as string) || "",
+          toolName: toolName || "",
+          state: part.state as ToolInvocationInfo["state"],
+          input: part.input,
+          output: part.output,
+        };
+      });
+  };
+
+  // Render message content from parts (no content property, only parts)
+  const renderMessageContent = (
+    messageParts: Array<{ type: string; text?: string }> | undefined,
+  ) => {
+    // Messages only have parts, no content property
+    if (messageParts && messageParts.length > 0) {
+      const textParts = messageParts.filter(p => p.type === "text" && p.text);
+      if (textParts.length > 0) {
+        return textParts.map((part, idx) => (
+          <ReactMarkdown
+            key={idx}
+            remarkPlugins={[remarkGfm]}
+            components={{
+              code({ className, children }) {
+                const match = /language-(\w+)/.exec(className || "");
+                const isInline = !match;
+                const codeString = String(children).replace(/\n$/, "");
+                return !isInline ? (
+                  <CodeBlock
+                    language={match ? match[1] : "text"}
+                    key={codeString}
+                    isGenerating={false}
+                  >
+                    {codeString}
+                  </CodeBlock>
+                ) : (
+                  <code className={className} style={{ fontSize: "0.8rem" }}>
+                    {children}
+                  </code>
+                );
+              },
+              table({ children }) {
+                return (
+                  <Box sx={{ overflow: "auto", my: 1 }}>
+                    <table
+                      style={{
+                        borderCollapse: "collapse",
+                        width: "100%",
+                        fontSize: "0.875rem",
+                        border: `1px solid ${muiTheme.palette.divider}`,
+                      }}
+                    >
+                      {children}
+                    </table>
+                  </Box>
+                );
+              },
+              th({ children }) {
+                return (
+                  <th
+                    style={{
+                      padding: "8px 12px",
+                      textAlign: "left",
+                      backgroundColor: muiTheme.palette.background.paper,
+                      borderBottom: `2px solid ${muiTheme.palette.divider}`,
+                      borderRight: `1px solid ${muiTheme.palette.divider}`,
+                      fontWeight: 600,
+                    }}
+                  >
+                    {children}
+                  </th>
+                );
+              },
+              td({ children }) {
+                return (
+                  <td
+                    style={{
+                      padding: "8px 12px",
+                      borderBottom: `1px solid ${muiTheme.palette.divider}`,
+                      borderRight: `1px solid ${muiTheme.palette.divider}`,
+                      backgroundColor: muiTheme.palette.background.paper,
+                    }}
+                  >
+                    {children}
+                  </td>
+                );
+              },
+            }}
+          >
+            {part.text || ""}
+          </ReactMarkdown>
+        ));
+      }
+    }
+
+    return null;
+  };
 
   return (
     <Box sx={{ display: "flex", flexDirection: "column", height: "100%" }}>
@@ -1056,22 +1045,6 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
             >
               Chat
             </Typography>
-            <Tooltip
-              title="Switch to AI SDK v3 (experimental)"
-              placement="bottom"
-            >
-              <Switch
-                size="small"
-                checked={useChatV3}
-                onChange={e => setUseChatV3(e.target.checked)}
-                sx={{ ml: 0.5 }}
-              />
-            </Tooltip>
-            {useChatV3 && (
-              <Typography variant="caption" color="text.secondary">
-                v3
-              </Typography>
-            )}
           </Box>
           <Box sx={{ display: "flex", alignItems: "center", gap: 0.5 }}>
             <IconButton size="small" onClick={createNewSession}>
@@ -1100,14 +1073,14 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
         {sessions
           .filter(
             session =>
-              session._id === sessionId ||
+              session._id === chatId ||
               (session.title && session.title.length > 0),
           )
           .map(session => (
             <MenuItem
               key={session._id}
               onClick={() => handleSelectSession(session._id)}
-              selected={session._id === sessionId}
+              selected={session._id === chatId}
               sx={{ display: "flex", justifyContent: "space-between" }}
             >
               <Box sx={{ display: "flex", alignItems: "center", flex: 1 }}>
@@ -1154,12 +1127,8 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
       {/* Error display */}
       {error && (
         <Box sx={{ p: 1 }}>
-          <Alert
-            severity="error"
-            onClose={() => setError(null)}
-            sx={{ fontSize: "0.875rem" }}
-          >
-            {error}
+          <Alert severity="error" sx={{ fontSize: "0.875rem" }}>
+            {error.message}
           </Alert>
         </Box>
       )}
@@ -1167,124 +1136,98 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
       {/* Messages */}
       <Box sx={{ flex: messages.length > 0 ? 1 : 0, overflow: "auto", p: 1 }}>
         <List dense>
-          {messages.map((m, idx) => (
-            <MessageItem
-              key={idx}
-              message={m}
-              isLastUser={idx === messages.length - 1 && m.role === "user"}
-              loading={loading}
-              onToolClick={handleToolClick}
-            />
-          ))}
-          {loading && (
-            <>
-              {/* Show streaming assistant response */}
-              <ListItem alignItems="flex-start" sx={{ p: 0 }}>
+          {messages.map(message => (
+            <ListItem key={message.id} alignItems="flex-start" sx={{ p: 0 }}>
+              {message.role === "user" ? (
+                <Box sx={{ flex: 1 }}>
+                  <Paper
+                    variant="outlined"
+                    sx={{
+                      p: 1,
+                      borderRadius: 1,
+                      backgroundColor: "background.paper",
+                      overflow: "hidden",
+                    }}
+                  >
+                    <Box
+                      sx={{
+                        overflow: "auto",
+                        maxWidth: "100%",
+                      }}
+                    >
+                      <ListItemText
+                        primary={
+                          // Extract text from parts
+                          (message.parts || [])
+                            .filter(
+                              (p): p is { type: "text"; text: string } =>
+                                p.type === "text" && "text" in p,
+                            )
+                            .map(p => p.text)
+                            .join("") || ""
+                        }
+                        primaryTypographyProps={{
+                          variant: "body2",
+                          color: "text.primary",
+                          sx: {
+                            whiteSpace: "pre-wrap",
+                            wordBreak: "break-word",
+                            overflowWrap: "break-word",
+                          },
+                        }}
+                      />
+                    </Box>
+                  </Paper>
+                </Box>
+              ) : (
                 <Box
                   sx={{
                     flex: 1,
                     overflow: "hidden",
                     fontSize: "0.875rem",
-                    "& pre": {
-                      margin: 0,
-                      overflow: "hidden",
-                    },
+                    "& pre": { margin: 0, overflow: "hidden" },
                   }}
                 >
-                  {/* Display tool calls first (chronologically before response) */}
+                  {/* Display tool invocations */}
                   <ToolCallsDisplay
-                    toolCalls={streamingToolCalls}
+                    toolInvocations={getToolInvocations(
+                      message.parts as Array<Record<string, unknown>>,
+                    )}
                     onToolClick={handleToolClick}
                   />
-
-                  {/* Then show the streaming content */}
-                  <ReactMarkdown
-                    remarkPlugins={[remarkGfm]}
-                    components={{
-                      code({ className, children }) {
-                        const match = /language-(\w+)/.exec(className || "");
-                        const isInline = !match;
-                        const codeString = String(children).replace(/\n$/, "");
-                        return !isInline ? (
-                          <CodeBlock
-                            language={match ? match[1] : "text"}
-                            key={codeString}
-                            isGenerating
-                          >
-                            {codeString}
-                          </CodeBlock>
-                        ) : (
-                          <code
-                            className={className}
-                            style={{ fontSize: "0.8rem" }}
-                          >
-                            {children}
-                          </code>
-                        );
-                      },
-                      table({ children }) {
-                        const muiTheme = useMuiTheme();
-                        return (
-                          <Box sx={{ overflow: "auto", my: 1 }}>
-                            <table
-                              style={{
-                                borderCollapse: "collapse",
-                                width: "100%",
-                                fontSize: "0.875rem",
-                                border: `1px solid ${muiTheme.palette.divider}`,
-                              }}
-                            >
-                              {children}
-                            </table>
-                          </Box>
-                        );
-                      },
-                      th({ children }) {
-                        const muiTheme = useMuiTheme();
-                        return (
-                          <th
-                            style={{
-                              padding: "8px 12px",
-                              textAlign: "left",
-                              backgroundColor:
-                                muiTheme.palette.background.paper,
-                              borderBottom: `2px solid ${muiTheme.palette.divider}`,
-                              borderRight: `1px solid ${muiTheme.palette.divider}`,
-                              fontWeight: 600,
-                            }}
-                          >
-                            {children}
-                          </th>
-                        );
-                      },
-                      td({ children }) {
-                        const muiTheme = useMuiTheme();
-                        return (
-                          <td
-                            style={{
-                              padding: "8px 12px",
-                              borderBottom: `1px solid ${muiTheme.palette.divider}`,
-                              borderRight: `1px solid ${muiTheme.palette.divider}`,
-                              backgroundColor:
-                                muiTheme.palette.background.paper,
-                            }}
-                          >
-                            {children}
-                          </td>
-                        );
-                      },
-                    }}
-                  >
-                    {streamingContent ||
-                      (steps.length === 0 && streamingToolCalls.length === 0
-                        ? "Assistant is thinking..."
-                        : "")}
-                  </ReactMarkdown>
+                  {/* Display reasoning */}
+                  <ReasoningDisplay
+                    messageParts={
+                      message.parts as Array<Record<string, unknown>>
+                    }
+                  />
+                  {/* Display message content */}
+                  {renderMessageContent(
+                    message.parts as Array<{ type: string; text?: string }>,
+                  )}
                 </Box>
-              </ListItem>
-            </>
+              )}
+            </ListItem>
+          ))}
+
+          {/* Loading indicator */}
+          {status === "submitted" && (
+            <ListItem alignItems="flex-start" sx={{ p: 0 }}>
+              <Box
+                sx={{
+                  flex: 1,
+                  overflow: "hidden",
+                  fontSize: "0.875rem",
+                }}
+              >
+                <Typography variant="body2" color="text.secondary">
+                  Assistant is thinking...
+                </Typography>
+              </Box>
+            </ListItem>
           )}
         </List>
+        <div ref={messagesEndRef} />
       </Box>
 
       {/* Input */}
@@ -1301,83 +1244,108 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
           gap: 1,
         }}
       >
-        {/* Text Area */}
-        <TextField
-          fullWidth
-          autoFocus
-          multiline
-          minRows={1}
-          maxRows={6}
-          placeholder="Ask Chat..."
-          value={input}
-          onChange={e => setInput(e.target.value)}
-          onKeyDown={e => {
-            if (e.key === "Enter" && !e.shiftKey) {
-              e.preventDefault();
-              sendMessage();
+        <form
+          onSubmit={e => {
+            e.preventDefault();
+            if (input.trim() && !isLoading) {
+              sendMessage({ text: input });
+              setInput("");
             }
-          }}
-          disabled={loading}
-          variant="outlined"
-          inputRef={inputRef}
-          sx={{
-            m: 0.5,
-            maxHeight: "50vh",
-            overflowY: "auto",
-            "& .MuiInputBase-input": {
-              fontSize: 14,
-            },
-            "& .MuiInputBase-root": {
-              p: 0,
-              fontSize: 14,
-            },
-            "& .MuiOutlinedInput-notchedOutline": {
-              border: "none",
-            },
-            "& .MuiOutlinedInput-root:hover .MuiOutlinedInput-notchedOutline": {
-              border: "none",
-            },
-            "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline":
-              {
-                border: "none",
-              },
-          }}
-        />
-
-        {/* Bottom action bar with Model Selector on left, Send button on right */}
-        <Box
-          sx={{
-            display: "flex",
-            justifyContent: "space-between",
-            alignItems: "center",
           }}
         >
-          {/* Model Selector */}
-          <Box sx={{ display: "flex", alignItems: "center" }}>
-            <ModelSelector />
-          </Box>
-
-          {/* Send Button */}
-          <IconButton
-            onClick={sendMessage}
-            disabled={
-              loading || !input.trim() || !sessionId || !currentWorkspace
-            }
-            size="small"
+          <TextField
+            fullWidth
+            autoFocus
+            multiline
+            minRows={1}
+            maxRows={6}
+            placeholder="Ask Chat..."
+            value={input}
+            onChange={e => setInput(e.target.value)}
+            onKeyDown={e => {
+              if (e.key === "Enter" && !e.shiftKey) {
+                e.preventDefault();
+                if (input.trim() && !isLoading) {
+                  sendMessage({ text: input });
+                  setInput("");
+                }
+              }
+            }}
+            disabled={isLoading}
+            variant="outlined"
+            inputRef={inputRef}
             sx={{
-              color:
-                input.trim() && currentWorkspace
-                  ? "primary.main"
-                  : "text.disabled",
-              p: 0,
-              "&:hover": {
-                backgroundColor: "action.hover",
+              m: 0.5,
+              maxHeight: "50vh",
+              overflowY: "auto",
+              "& .MuiInputBase-input": {
+                fontSize: 14,
               },
+              "& .MuiInputBase-root": {
+                p: 0,
+                fontSize: 14,
+              },
+              "& .MuiOutlinedInput-notchedOutline": {
+                border: "none",
+              },
+              "& .MuiOutlinedInput-root:hover .MuiOutlinedInput-notchedOutline":
+                {
+                  border: "none",
+                },
+              "& .MuiOutlinedInput-root.Mui-focused .MuiOutlinedInput-notchedOutline":
+                {
+                  border: "none",
+                },
+            }}
+          />
+
+          {/* Bottom action bar with Model Selector on left, Send/Stop button on right */}
+          <Box
+            sx={{
+              display: "flex",
+              justifyContent: "space-between",
+              alignItems: "center",
             }}
           >
-            <SendIcon sx={{ fontSize: 20 }} />
-          </IconButton>
-        </Box>
+            <Box sx={{ display: "flex", alignItems: "center" }}>
+              <ModelSelector />
+            </Box>
+
+            {isLoading ? (
+              <IconButton
+                onClick={stop}
+                size="small"
+                sx={{
+                  color: "error.main",
+                  p: 0,
+                  "&:hover": {
+                    backgroundColor: "action.hover",
+                  },
+                }}
+              >
+                <StopIcon sx={{ fontSize: 20 }} />
+              </IconButton>
+            ) : (
+              <IconButton
+                type="submit"
+                disabled={!input.trim() || !currentWorkspace}
+                size="small"
+                sx={{
+                  color:
+                    input.trim() && currentWorkspace
+                      ? "primary.main"
+                      : "text.disabled",
+                  p: 0,
+                  "&:hover": {
+                    backgroundColor: "action.hover",
+                  },
+                }}
+              >
+                <SendIcon sx={{ fontSize: 20 }} />
+              </IconButton>
+            )}
+          </Box>
+        </form>
       </Paper>
 
       {/* Tool Debug Dialog */}
@@ -1388,9 +1356,7 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
         maxWidth="md"
       >
         <DialogTitle>
-          {selectedToolCall
-            ? `Tool: ${selectedToolCall.toolName}`
-            : "Tool Details"}
+          {selectedTool ? `Tool: ${selectedTool.toolName}` : "Tool Details"}
         </DialogTitle>
         <DialogContent dividers>
           <Box sx={{ mb: 2 }}>
@@ -1398,10 +1364,10 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
               Input
             </Typography>
             <CodeBlock language="json" isGenerating={false} scrollable>
-              {selectedToolCall && selectedToolCall.input !== undefined
-                ? typeof selectedToolCall.input === "string"
-                  ? selectedToolCall.input
-                  : JSON.stringify(selectedToolCall.input, null, 2)
+              {selectedTool && selectedTool.input !== undefined
+                ? typeof selectedTool.input === "string"
+                  ? selectedTool.input
+                  : JSON.stringify(selectedTool.input, null, 2)
                 : "No input captured"}
             </CodeBlock>
           </Box>
@@ -1410,10 +1376,10 @@ const Chat: React.FC<ChatProps> = ({ onConsoleModification }) => {
               Output
             </Typography>
             <CodeBlock language="json" isGenerating={false} scrollable>
-              {selectedToolCall && selectedToolCall.result !== undefined
-                ? typeof selectedToolCall.result === "string"
-                  ? selectedToolCall.result
-                  : JSON.stringify(selectedToolCall.result, null, 2)
+              {selectedTool && selectedTool.output !== undefined
+                ? typeof selectedTool.output === "string"
+                  ? selectedTool.output
+                  : JSON.stringify(selectedTool.output, null, 2)
                 : "No output captured"}
             </CodeBlock>
           </Box>
